requirements.txt
Flask
mysql-connector-python
waitress
email-validator